fully working SDK

+++ wandb/sdk/internal/internal_api.py
@@ -3151,7 +3151,7 @@ class Api:
             $labels: JSONString,
             $aliases: [ArtifactAliasInput!],
             $metadata: JSONString,
-            $ttlDurationSeconds: int,
+            $ttlDurationSeconds: Int64,
             {}
             {}
             {}
